Create CAT P2 Q5
Write a C Program that reads the input as a string from the user in main (). (1 
mark)
The input should be a sentence with two words. Pass this string to a function.(1 
mark) 
Inside the function do the following operations:  For the first word, keep only the first character of the word to be in upper 
case and the rest of the characters in lower case. (1 mark)  For the second word, convert all the characters into upper case letter. (1 
mark)  Print the revised string consisting of the two words in the function itself 
(1 mark)  Find the length of the entire string. Print its length in the function itself in 
the next line of the revised string. Use appropriate string function to print 
this result in the next line. (1 mark)  Return the length of the string, if the length is less than 20. Else return the 
size of the string. (2 marks)
Consider the input string given by the user is:  “computer programming”  Revised string to be printed in the function is “Computer 
PROGRAMMING”

The length of a string is the number of characters in the string, excluding the null terminator,
 while the size of a string is the total number of bytes allocated for the string, including the null terminator.
In other words, the size includes the length plus one for the null terminator.

For the input string "computer programming", the revised string to be printed in the function is "Computer PROGRAMMING".
The length of this string is 20, so the function will return the size of the string, which is also 20.